# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "post": {
        "description": "Creates a group.      The user is automatically added as an owner.      The owner must not be among the members.",
        "operationId": "GroupController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "description": "Lists all groups.      Only groups that the user is a member of are returned.",
        "operationId": "GroupController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}": {
      "get": {
        "operationId": "GroupController_findOne",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Groups"
        ]
      },
      "patch": {
        "description": "Partially updates a group.      Only the provided fields are updated. The rest remain the same.",
        "operationId": "GroupController_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "description": "Deletes a group.      Only the owner can delete a group.      The group is not actually deleted but marked as deleted.",
        "operationId": "GroupController_deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/invite": {
      "post": {
        "operationId": "GroupController_addMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/v1/groups/{groupId}/members/{userId}": {
      "delete": {
        "operationId": "GroupController_removeMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Groups"
        ]
      },
      "patch": {
        "operationId": "GroupController_updateMember",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/api/v1/ledgers/{ledgerId}/expenses": {
      "post": {
        "operationId": "ExpensesController_createExpense",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Expenses"
        ]
      },
      "get": {
        "operationId": "ExpensesController_listExpenses",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Expenses"
        ]
      }
    },
    "/api/v1/expenses/{expenseId}": {
      "patch": {
        "operationId": "ExpensesController_updateExpense",
        "parameters": [
          {
            "name": "expenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Expenses"
        ]
      },
      "delete": {
        "operationId": "ExpensesController_voidExpense",
        "parameters": [
          {
            "name": "expenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Expenses"
        ]
      }
    },
...
    "/api/v1/ledgers/{ledgerId}/balances": {
      "get": {
        "operationId": "BalancesController_getBalances",
        "parameters": [
          {
            "name": "ledgerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Balances"
        ]
      }
    }
  },
  "info": {
    "title": "Expensive APIs",
    "description": "The API description for the Expensive app.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserPublicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "fullName"
        ]
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "GroupRole": {
        "type": "string",
        "enum": [
          "OWNER",
          "ADMIN",
          "MEMBER"
        ]
      },
      "GroupMembershipStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "REMOVED"
        ]
      },
      "GroupMemberReducedDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupMembershipStatus"
              }
            ]
          },
          "user": {
            "$ref": "#/components/schemas/UserPublicDTO"
          }
        },
        "required": [
          "userId",
          "role",
          "status",
          "user"
        ]
      },
      "GroupDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdById": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberReducedDTO"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdById",
          "createdAt"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object"
          },
          "description": {
            "type": "object"
          }
        }
      },
      "AddGroupMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              }
            ]
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateGroupMemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupRole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroupMembershipStatus"
              }
            ]
          }
        }
      },
...
      "SplitMethod": {
        "type": "string",
        "enum": [
          "EQUAL",
          "EXACT",
          "PERCENTAGE",
          "ITEMIZED"
        ]
      },
      "ExpenseSplitDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "amountOwed": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "amountOwed"
        ]
      },
      "ExpenseItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "amount",
          "quantity"
        ]
      },
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "payerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "totalAmount": {
            "type": "string"
          },
          "expenseDate": {
            "type": "string"
          },
          "splitMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SplitMethod"
              }
            ]
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseSplitDto"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseItemDto"
            }
          }
        },
        "required": [
          "payerId",
          "title",
          "totalAmount",
          "expenseDate",
          "splitMethod",
          "splits"
        ]
      },
      "ExpenseResponseSplitDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "amountOwed": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "amountOwed"
        ]
      },
      "ExpenseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ledgerId": {
            "type": "string"
          },
          "payerId": {
            "type": "string"
          },
          "payer": {
            "$ref": "#/components/schemas/UserPublicDTO"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "string"
          },
          "expenseDate": {
            "type": "string",
            "format": "date-time"
          },
          "splitMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SplitMethod"
              }
            ]
          },
          "isVoided": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseResponseSplitDto"
            }
          }
        },
        "required": [
          "id",
          "ledgerId",
          "payerId",
          "payer",
          "title",
          "totalAmount",
          "expenseDate",
          "splitMethod",
          "isVoided",
          "createdAt",
          "splits"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "expenseDate": {
            "type": "string"
          },
          "splitMethod": {
            "type": "string"
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseSplitDto"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseItemDto"
            }
          }
        }
      },
...
    }
  }
}